Ensure a simple 'make' command works when starting out.
Before looking deeply and figuring it out, building this from
scratch by typing 'make' ends up failing in the ostree update
because there is no 'repo' found.

Fix this by simply making the 'repo' a prerequisite of updating
the ostree repository. This will only create the repository once
if the repo directory is not found.