cross-localedef-native: Fix build on aarch64
This adds an override cross-localedef-native_2.22.bbappend override
which leverages the oe autotools configure, ultimately so that
it will use a fresh config.sub & config.guess and properly recognize
the aarch64 machine.

This is a backport of an upstream yocto bug filed with this report
and patch mail:

  https://bugzilla.yoctoproject.org/show_bug.cgi?id=9508
  http://lists.openembedded.org/pipermail/openembedded-core/2016-April/120613.html

This should be removed if/when the upstream bug is fixed in jethro, or, when
moving to a new poky release branch which uses a newer version of glibc.