Remove stale .py[co] files in base layer
This caused problems in the higher levels. flatpak-builder now does
mtime rewriting in during commits to handle ostree resetting the mtime
during commit/checkout.

In order for this to work flatpak-builder assumes that what is in the
base layer is correct. i.e. if the base layer has a .py + a .pyc file
both with now cleared mtimes, but the embedded mtime of the .pyc file
is not cleared, then we clear it, assuming it is not stale. This is
correct since this is mostly true for the base layer, and we want to
have such .pyc files for the core python lib.

However, if the base layer has a stale .pyc file then this will
propagate that to a non-stale one, causing problems.

To fix this we remove all stale .pyc files with a simple mtime check
before committing. For python 3.x the check is more complex, so we
just remove all __pycache__ dirs (we don't have any atm anyway).