change GCC / binutils architecture triple on ARM
Despite fixing the actual compiler flags in the previous commit, the
remaining problem is that Yocto doesn't understand/use the correct
architecture triple for armhf, namely arm-linux-gnueabihf, because it
instead sets all of the build options, CFLAGS, etc in each recipe. This
means the compiler that Yocto uses, and produces, self-identifies as
arm-unknown-linux-gnueabi (ie, armel) even though it is only ever used
in a mode that generates hard float binaries. However, when running
configure scripts outside of Yocto, or even just examining the compiler,
it's deeply confusing to have an armhf compiler which claims to be
armel.

After winning a fight to the death with the Yocto build system, based
off a WIP from Richard Purdie, this changes GCC and binutils to build
and install using the correct architecture target, arm-linux-gnueabihf.

Based off upstream
http://git.yoctoproject.org/cgit.cgi/poky-contrib/commit/?h=rpurdie/wip5&id=b0c39d3e72fc12a5c8cbdd9769c1baf1b6a5b23f